requirements.txt
tensorflow==1.13.2
matplotlib
numpy
csv

tabulate
imageio
pillow
